Extract heatmap and histogram code into separate modules
